Move Cells to new ViewComponents
